@@ -1,26 +1,82 @@
 [tool.poetry]
-name = "python-pinballmap"
+name = "pinballmap"
 version = "0.3.8"
 description = "Python client for the Pinball Map API"
 authors = ["Duna Csandl <marinas.bobble-05@icloud.com>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = ">=3.7,<4.0.0"
+python = ">=3.11,<4.0.0"
+requests = "^2.32.3"
+tabulate = "^0.9.0"
 
-requests = "^2.27.1"
-tabulate = "^0.8.9"
 
-[tool.poetry.dev-dependencies]
-black = "^21.12b0"
-bump2version = "^1.0.1"
-ipython = "^7"
+[tool.poetry.group.dev.dependencies]
+bump2version = "*"
+ipython = "^8.26.0"
 jupyter = "^1.0.0"
-pytest = "^6.2.5"
-Sphinx = "^4.3.2"
-twine = "^3.7.1"
-readme-renderer = "^32.0"
+pytest = "^8.3.2"
+Sphinx = "<8.0.0"
+twine = "*"
+readme-renderer = "*"
+ruff = "^0.5.5"
+isort = "^5.13.2"
+sphinx-rtd-theme = "^2.0.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+pinballmap = "pinballmap.cli:cli"
+
+[tool.isort]
+profile = "black"
+
+[tool.pytest]
+log_cli_level = "info"
+
+[tool.ruff]
+extend-exclude = ["docs/yourapp"]
+line-length = 88
+target-version = "py312"
+
+
+[tool.ruff.lint]
+select = [
+    # pycodestyle
+    "E",
+    # Pyflakes
+    "F",
+    # pyupgrade
+    "UP",
+    # flake8-bugbear
+    "B",
+    # flake8-simplify
+    "SIM",
+    # isort
+    "I",
+]
+ignore = ["RUF002", "COM812"]
+
+# Allow fix for all enabled rules (when `--fix`) is provided.
+fixable = ["ALL"]
+unfixable = []
+
+# Allow unused variables when underscore-prefixed.
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+docstring-code-format = true
+docstring-code-line-length = 60
+
+
+[tool.coverage.run]
+omit = [
+    "tests/*",
+]
