@@ -1,8 +1,6 @@
-# coding: utf-8
-
 import logging
+from collections.abc import Iterable
 from operator import itemgetter
-from typing import Dict, Iterable, List, Tuple, Union
 
 import requests
 
@@ -26,22 +24,27 @@ logger = logging.getLogger(__name__)
 
 class PinballMapClient:
     """
-    Creates a ``PinballMapClient``, optionally locked to a specific location_id and region name.
-    It will use Django's default cache if installed and available.
+    Creates a ``PinballMapClient``, optionally locked to a specific location_id and
+    region name. It will use Django's default cache if installed and available.
 
-    Passed-in values are optional. Authentication Token and email is needed for all write operations.
+    Passed-in values are optional. Authentication Token and email is needed for all
+    write operations.
 
-    You can use an instance to look up machines by name or id without specifying a location or region.
-    Most other methods will fail without them.
+    You can use an instance to look up machines by name or id without specifying a
+    location or region. Most other methods will fail without them.
 
-    If Django is installed and is in ``DEBUG`` mode, any write operations to the API will operate as a
-    "dry run". This is to prevent you from accidentally updating the map with inaccurate development data.
-    (You're welcome.)
+    If Django is installed and is in ``DEBUG`` mode, any write operations to the API
+    will operate as a "dry run". This is to prevent you from accidentally updating the
+    map with inaccurate development data. (You're welcome.)
 
-    Note: the API uses names a bit inconsistently… sometimes it's user_token, sometimes
-    it's authentication_token. Sometimes username, sometimes login. Always check the docs when adding/changing code here.
+    .. note::
 
-    :param authentication_token: Your Pinball Map API Authentication Token, needed for all write operations.
+        The API uses names a bit inconsistently… sometimes it's user_token, sometimes
+        it's authentication_token. Sometimes username, sometimes login. Always check
+        the docs when adding/changing code here.
+
+    :param authentication_token: Your Pinball Map API Authentication Token, needed for
+                                 all write operations.
     :param user_email: map account email
     :param user_password: map account password
     :param location_id: Your location_id, as found in the Pinball Map data
@@ -84,96 +87,102 @@ class PinballMapClient:
                 self.cache_key_prefix = settings.PINBALL_MAP.get(
                     "cache_key_prefix", self.cache_key_prefix
                 )
-                self.dry_run = True if settings.DEBUG is True else False
+                self.dry_run = settings.DEBUG is True
+
             except Exception as exc:
                 raise ValueError(
-                    "Could not use your Django settings for Pinball Map API because: {}".format(
-                        exc
-                    )
-                )
+                    f"Could not use your Django settings for Pinball Map API "
+                    f"because: {exc}"
+                ) from exc
+
         if caches:
             self.cache = caches[self.cache_name]
+
         self.lmxs = []
         self.all_machines = []
         self.session = requests.Session()
         if not self.authentication_token and self.user_email and self.user_password:
-            # attempt to get token from email and password, fail quietly so it's possible to try again:
+            # attempt to get token from email and password, fail quietly so it's
+            # possible to try again:
             details = dict()
             try:
                 details = self.auth_details(self.user_email, self.user_password)
+
             except Exception as msg:
-                logger.warning(
-                    "PinballMapClient self-authentication failed. {}".format(msg)
-                )
+                logger.warning(f"PinballMapClient self-authentication failed. {msg}")
+
             if "authentication_token" in details:
                 self.authentication_token = details["authentication_token"]
+
         if not self.user_email or not self.authentication_token:
             logger.info(
-                "Without user_email and authentication_token, all write operations will fail."
+                "Without user_email and authentication_token, all write operations "
+                "will fail."
             )
 
-    def get_all_machines(self) -> List[Dict]:
+    def get_all_machines(self) -> list[dict]:
         """
         Get list of all machines from PM DB. Cached to avoid a zillion large requests.
-        Currently we are not storing results in the object because it's unlikely for an instance to be re-used
-        for multiple name searches.
+        Currently we are not storing results in the object because it's unlikely for an
+        instance to be re-used for multiple name searches.
 
         :return: list of every machine
         """
-        cache_key = "{}_machines".format(self.cache_key_prefix)
+        cache_key = f"{self.cache_key_prefix}_machines"
         if self.cache:
             data = self.cache.get(cache_key, {})
             if data:
                 return data
-        r = self.session.get("{}/machines.json".format(self.BASE_URL))
+
+        r = self.session.get(f"{self.BASE_URL}/machines.json")
         if r.status_code != requests.codes.ok:
             logger.error(
-                "Getting list of all pinball map games failed with status code {}".format(
-                    r.status_code
-                )
+                f"Getting list of all pinball map games failed with status "
+                f"code {r.status_code}"
             )
             r.raise_for_status()
+
         data = r.json()["machines"]
+
         # patch raw data with searchable cleaned names
         for g in data:
             g["cleaned_name"] = clean_name(g["name"])
+
         if self.cache:
             self.cache.set(cache_key, data, 15 * 60)
+
         return data
 
-    def get_location_machine_xrefs(self) -> List[Dict]:
+    def get_location_machine_xrefs(self) -> list[dict]:
         """
-        Gets the list of location_machine_xrefs (LMXs) for our region, then filter it to include only ones from
-        our location_id.
+        Gets the list of location_machine_xrefs (LMXs) for our region, then filter
+        it to include only ones from our location_id.
 
-        Since this API request is massive and slow, and likely to be accessed multiple times in the lifetime on an
-        instance (when syncing, for example), we are using two levels of caching here.
-        We store the filtered LMXs for our location in the PinballMapClient instance (in-memory),
-        and ALSO cache the results for 15 minutes in Django's cache
-        (if available). We are assuming no instance of this object will ever live long enough to be
-        concerned about needing to bust its own copy of the cached data, but it will benefit from recent previous
-        runs.
+        Since this API request is massive and slow, and likely to be accessed multiple
+        times in the lifetime on an instance (when syncing, for example), we are using
+        two levels of caching here. We store the filtered LMXs for our location in the
+        PinballMapClient instance (in-memory), and ALSO cache the results for 15 minutes
+        in Django's cache (if available). We are assuming no instance of this object
+        will ever live long enough to be concerned about needing to bust its own copy
+        of the cached data, but it will benefit from recent previous runs.
 
         :return: list of LMXs for ``location_id``
         """
         if len(self.lmxs) > 0:
             return self.lmxs
-        cache_key = "{}_lmxs".format(self.cache_key_prefix)
+        cache_key = f"{self.cache_key_prefix}_lmxs"
         if self.cache:
             data = self.cache.get(cache_key, [])
             if data:
                 self.lmxs = data
                 return data
             del data
-        url = "{BASE_URL}/region/{region_name}/location_machine_xrefs.json".format(
-            BASE_URL=self.BASE_URL, region_name=self.region_name
-        )
+        url = f"{self.BASE_URL}/region/{self.region_name}/location_machine_xrefs.json"
         r = self.session.get(url)
         if r.status_code != requests.codes.ok:
             logger.error(
-                "Getting list of all pinball map games failed with status code {}".format(
-                    r.status_code
-                )
+                f"Getting list of all pinball map games failed with status "
+                f"code {r.status_code}"
             )
         r.raise_for_status()
         raw_data = r.json()["location_machine_xrefs"]
@@ -188,15 +197,16 @@ class PinballMapClient:
 
     def machine_by_name(
         self, query_string: str, min_score: int = 2, include_score: bool = False
-    ) -> Union[Tuple[Dict], Tuple[Tuple[Dict, str]]]:
+    ) -> tuple[dict] | tuple[tuple[dict, str]]:
         """
-        Finds likely name matches from the Pinball Map database and sorts results by a match quality score.
+        Finds likely name matches from the Pinball Map database and sorts results by a
+        match quality score.
 
         :param query_string: name of the game
         :param min_score: minimum quality score for matches. Our default of 2 seems to be the sweet spot.
         :param include_score: whether to include the match quality scores. Default = ``False``.
         :return: matches
-        """
+        """  # noqa: E501
         all_games = self.get_all_machines()
         query_string = clean_name(query_string)
         query_words = tuple(query_string.split())
@@ -210,15 +220,19 @@ class PinballMapClient:
                 results.append((data, score))
                 if score >= 150:
                     set_high_bar = True
+
         final_results = tuple(sorted(results, key=itemgetter(1), reverse=True))
         if set_high_bar:
-            # if we ever hit a high match score, only include top 5. At that point there is no need to grab at straws.
+            # if we ever hit a high match score, only include top 5. At that point there
+            # is no need to grab at straws.
             final_results = tuple(final_results[:4])
+
         if include_score:
             return final_results
+
         return tuple(map(itemgetter(0), final_results))
 
-    def machine_by_ipdb_id(self, ipdb_id: int) -> Union[Dict, None]:
+    def machine_by_ipdb_id(self, ipdb_id: int) -> dict | None:
         """
         Find machine by IPDB number
 
@@ -229,9 +243,10 @@ class PinballMapClient:
         for g in all_games:
             if g["ipdb_id"] == ipdb_id:
                 return g
+
         return None
 
-    def machine_by_map_id(self, map_id: int) -> Union[Dict, None]:
+    def machine_by_map_id(self, map_id: int) -> dict | None:
         """
         Find machine by pinball map ID
 
@@ -242,41 +257,43 @@ class PinballMapClient:
         for g in all_games:
             if g["id"] == map_id:
                 return g
+
         return None
 
-    def machines_at_location(self, location_id: int = None) -> List[Dict]:
+    def machines_at_location(self, location_id: int = None) -> list[dict]:
         """
         List the machines at location_id
 
-        :param location_id: optional location_id, or it will use the one in settings or set at init
+        :param location_id: optional location_id, or it will use the one in settings or
+                            set at init
         :return: list of machines matching location_id
         """
         if not location_id:
             location_id = self.location_id
+
         if not location_id:
             raise ValueError("Need a location id")
+
         r = self.session.get(
-            "{base}/locations/{id}/machine_details.json".format(
-                base=self.BASE_URL, id=location_id
-            )
+            f"{self.BASE_URL}/locations/{location_id}/machine_details.json"
         )
         if r.status_code != requests.codes.ok:
             logger.error(
-                "Getting list of machines at our location failed with status code {}".format(
-                    r.status_code
-                )
+                f"Getting list of machines at our location failed with status "
+                f"code {r.status_code}"
             )
+
         r.raise_for_status()
         return r.json()["machines"]
 
-    def compare_location(self, my_machine_ids: Iterable[int]) -> Dict:
+    def compare_location(self, my_machine_ids: Iterable[int]) -> dict:
         """
-        Compares a machine list with Pinball Map's data. Returns a ``dict`` with which ids to
-        add, remove, or ignore (meaning they are already listed).
+        Compares a machine list with Pinball Map's data. Returns a ``dict`` with which
+        ids to add, remove, or ignore (meaning they are already listed).
 
         :param my_machine_ids: iterable of Pinball Map machine_ids
         :return: {'add': [id0, id1, idn...], 'remove': [id0, id1, idn...], 'ignore': [...]}
-        """
+        """  # noqa: E501
         my_machine_ids = frozenset(my_machine_ids)
         map_machine_ids = frozenset([g["id"] for g in self.machines_at_location()])
         add = my_machine_ids.difference(map_machine_ids)
@@ -284,7 +301,7 @@ class PinballMapClient:
         ignore = my_machine_ids.intersection(map_machine_ids)
         return dict(add=add, remove=remove, ignore=ignore)
 
-    def lmx_by_machine_id(self, machine_id: int) -> Dict:
+    def lmx_by_machine_id(self, machine_id: int) -> dict:
         """
         Get the ``location_machine_xref`` for ``machine_id`` at my location.
 
@@ -295,14 +312,18 @@ class PinballMapClient:
         for lmx in lmxs:
             if lmx["machine"]["id"] == machine_id:
                 return lmx
+
         return {}
 
     @requires_authorization
-    def remove_machine(self, machine_id: int) -> Union[Dict, None]:
+    def remove_machine(self, machine_id: int) -> dict | None:
         """
         Remove a machine from my location.
 
-        NOTE: If it detects that Django is in debug mode, it will not actually perform the removal.
+        .. note::
+
+            If it detects that Django is in debug mode, it will not actually
+            perform the removal.
 
         :param machine_id: machine id to remove
         :return: JSON result of operation (as a dict) or None
@@ -310,20 +331,19 @@ class PinballMapClient:
         lmx = self.lmx_by_machine_id(machine_id)
         if not lmx:
             logger.warning(
-                "Tried to remove machine_id {} from pinball map, but there was no matching LMX.".format(
-                    machine_id
-                )
+                f"Tried to remove machine_id {machine_id} from pinball map, but there "
+                f"was no matching LMX."
             )
             return None
+
         lmx_id = lmx["id"]
-        url = "{BASE_URL}/location_machine_xrefs/{lmx_id}.json".format(
-            BASE_URL=self.BASE_URL, lmx_id=lmx_id
-        )
+        url = f"{self.BASE_URL}/location_machine_xrefs/{lmx_id}.json"
         if self.dry_run:
             logger.warning(
-                "since Django is in DEBUG mode, I'm not going to DELETE {}".format(url)
+                f"since Django is in DEBUG mode, I'm not going to DELETE {url}"
             )
             return None
+
         params = {
             "user_email": self.user_email,
             "user_token": self.authentication_token,
@@ -332,30 +352,33 @@ class PinballMapClient:
         r = self.session.delete(url, params=params)
         if not ok_response_code(r.status_code):
             logger.error(
-                "Failed to remove id {} at URL {} status code={}, content: {}".format(
-                    machine_id, r.url, r.status_code, r.content
-                )
+                f"Failed to remove id {machine_id} at URL {r.url} "
+                f"status code={r.status_code}, content: {r.content}"
             )
+
         r.raise_for_status()
         result = r.json()
         return result
 
     @requires_authorization
-    def add_machine(self, machine_id: int) -> Union[Dict, None]:
+    def add_machine(self, machine_id: int) -> dict | None:
         """
         Add a machine to my location.
 
-        NOTE: If it detects that Django is in debug mode, it will not actually perform the addition.
+        .. note::
+            If it detects that Django is in debug mode, it will not actually perform
+            the addition.
 
         :param machine_id:
         :return: JSON result or None.
         """
-        url = "{BASE_URL}/location_machine_xrefs.json".format(BASE_URL=self.BASE_URL)
+        url = f"{self.BASE_URL}/location_machine_xrefs.json"
         if self.dry_run:
             logger.warning(
-                "since Django is in DEBUG mode, I'm not going to POST to {}".format(url)
+                f"since Django is in DEBUG mode, I'm not going to POST to {url}"
             )
             return None
+
         params = dict(
             user_email=self.user_email,
             user_token=self.authentication_token,
@@ -365,19 +388,19 @@ class PinballMapClient:
         r = self.session.post(url, params=params)
         if not ok_response_code(r.status_code):
             logger.warning(
-                "Failed to add id {} to Pinball Map at URL {} :: status code={}, content: {}".format(
-                    machine_id, r.url, r.status_code, r.content
-                )
+                f"Failed to add id {machine_id} to Pinball Map at URL {r.url} :: "
+                f"status code={r.status_code}, content: {r.content}"
             )
+
         r.raise_for_status()
         result = r.json()
         return result
 
     @requires_authorization
-    def update_map(self, machine_ids: Iterable[int]) -> Dict:
+    def update_map(self, machine_ids: Iterable[int]) -> dict:
         """
-        Given a complete list of machine_ids for the location, this will add and remove them as needed so that
-        Pinball Map matches your current list of machines.
+        Given a complete list of machine_ids for the location, this will add and remove
+        them as needed so that Pinball Map matches your current list of machines.
 
         :param machine_ids: the pinball map id numbers for your current list of machines
         :return: dict of count of machines added, removed, or ignored
@@ -389,17 +412,23 @@ class PinballMapClient:
         for machine_id in change_data["add"]:
             try:
                 self.add_machine(machine_id)
+
             except Exception as exc:
-                errors[machine_id] = "Failed to add: {}".format(exc)
+                errors[machine_id] = f"Failed to add: {exc}"
                 continue
+
             added.append(machine_id)
+
         for machine_id in change_data["remove"]:
             try:
                 self.remove_machine(machine_id)
+
             except Exception as exc:
-                errors[machine_id] = "Failed to remove: {}".format(exc)
+                errors[machine_id] = f"Failed to remove: {exc}"
                 continue
+
             removed.append(machine_id)
+
         return dict(
             added=len(added),
             removed=len(removed),
@@ -411,6 +440,7 @@ class PinballMapClient:
     def _update_self(self, data):
         if "errors" in data or "user" not in data:
             return
+
         user = data["user"]
         self.authentication_token = user["authentication_token"]
         self.user_email = user["email"]
@@ -422,14 +452,18 @@ class PinballMapClient:
         Creates a user account. Note: This is an easy way to get a token to use later.
         If login is successful, optionally set the token for this client.
 
-        If signup fails, attempt to log in and self-update with the same credentials in case the reason for failure is simply
-        that the account already exists.
+        If signup fails, attempt to log in and self-update with the same credentials in
+        case the reason for failure is simply that the account already exists.
 
-        NOTE: the API server will send an email with a verification link. You can't do anything further until confirmed.
+
+        .. note::
+            The API server will send an email with a verification link. You can't do
+            anything further until you confirm it.
 
         Example responses:
 
         .. code-block::json
+
             {
                 "user": {
                     "id": 1,
@@ -441,6 +475,7 @@ class PinballMapClient:
 
 
         .. code-block: json
+
             {
                 "errors": "your entered passwords do not match"
             }
@@ -449,20 +484,22 @@ class PinballMapClient:
         :param username: the username
         :param email: user's email address
         :param password: the password
-        :param update_self: whether to update this client instance's authentication_token, default is True
+        :param update_self: whether to update this client instance's
+                            authentication_token, default is True
         :return: result of auth_details request as dict
         """
         params = dict(
             username=username, email=email, password=password, confirm_password=password
         )
-        url = "{BASE_URL}/users/signup.json".format(BASE_URL=self.BASE_URL)
+        url = f"{self.BASE_URL}/users/signup.json"
         r = self.session.post(url, params=params)
         if not ok_response_code(r.status_code):
             logger.warning(
-                "Failed to create Pinball Map user account for {}. status code={}, "
-                "content={}".format(username, r.status_code, r.content)
+                f"Failed to create Pinball Map user account for {username}. "
+                f"status code={r.status_code}, content={r.content}"
             )
             r.raise_for_status()
+
         result = r.json()
         if "errors" in result:
             logger.error(
@@ -473,15 +510,16 @@ class PinballMapClient:
             # be persistent in case account exists by trying to log in anyway:
             try:
                 return self.auth_details(email, password, update_self=update_self)
+
             except PinballMapAuthenticationFailure:
                 logger.error(
-                    "Could neither create nor log into Pinball Map account {}".format(
-                        email
-                    )
+                    f"Could neither create nor log into Pinball Map account {email}"
                 )
                 raise
+
         if update_self:
             self._update_self(result)
+
         return result
 
     def auth_details(
@@ -489,20 +527,24 @@ class PinballMapClient:
     ) -> dict:
         """
         Gets the authorization details for user.
-        If login is successful, optionally set the token for this client (default is True).
+        If login is successful, optionally set the token for this client
+        (default is True).
 
-        If we have settings in Django, we'll use those automatically if username and password are None.
+        If we have settings in Django, we'll use those automatically if username and
+        password are None.
 
         Example responses:
         ------------------
 
         .. code-block:: json
+        
             {
                 "errors": "User is not yet confirmed. Please follow emailed confirmation instructions."
             }
 
 
         .. code-block:: json
+        
             {
                 "user": {
                     "id": 1,
@@ -515,26 +557,30 @@ class PinballMapClient:
 
         :param username: the username or email address
         :param password: the password
-        :param update_self: whether to update this client instance's authentication_token, default is True
+        :param update_self: whether to update this client instance's
+                            authentication_token, default is True
         :return: result of auth_details request as dict
-        """
+        """  # noqa: E501
         if username is None and password is None:
             raise ValueError("username and password required")
         params = dict(login=username, password=password)
-        url = "{BASE_URL}/users/auth_details.json".format(BASE_URL=self.BASE_URL)
+        url = f"{self.BASE_URL}/users/auth_details.json"
         r = self.session.get(url, params=params)
         if r.status_code != requests.codes.ok:
             logger.warning(
-                "Failed to get Pinball Map auth details for username: {}. status code={}, "
-                "content={}".format(username, r.status_code, r.content)
+                f"Failed to get Pinball Map auth details for username: {username}. "
+                f"status code={r.status_code}, content={r.content}"
             )
             r.raise_for_status()
+
         result = r.json()
         if "errors" in result:
             logger.error(
-                "username {} failed to authenticate to Pinball Map API".format(username)
+                f"username {username} failed to authenticate to Pinball Map API"
             )
             raise PinballMapAuthenticationFailure(str(result["errors"]))
+
         if update_self:
             self._update_self(result)
+
         return result
